+++ MicrosoftEntraID/CHANGELOG.md
@@ -7,6 +7,18 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## Unreleased
 
+## 2025-10-10 - 2.10.0
+
+### Added
+
+- Microsoft Entra ID Graph Api Connector
+
+## 2025-09-19 - 2.9.1
+
+### Fixed
+
+- Update Sekoia SDK and Asset configuration
+
 ## 2025-09-19 - 2.9.0
 
 ### Fixed